rename labchart2neo
move all the complicated plotting to a separate file. Only all the computation and simple plotting should be in this file. In the end all the results are saved to a pickle which can then be used in the plotting script. This also avoids that the big files have to be loaded into memory all the time..