fix the onset extraction bug, close #13
The problem was that I took the wrong onsets, e.g. the one when the stepper was signal was over instead of the one where the stepper started to be active. Furthermore the onsets where shifted by kernel_width/2 due to the convolution. Now this value is added, subtracted to lead the right values.